feat(gosub_winit): extract winit+wgpu presenter glue
